backend/requirements-simple.txt
fastapi==0.104.1
uvicorn==0.24.0
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
